+++ src/components/selectboxes/fixtures/index.js
@@ -1,4 +1,6 @@
 import { component as comp1 } from './comp1';
+import { component as comp2 } from './comp2';
 export const components = {
-  comp1: comp1
+  comp1: comp1,
+  comp2: comp2
 };